Combine commands and processors
