--- PerformanceCalculator/CommandBase.cs
@@ -1,12 +1,0 @@
-﻿// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
-// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-tools/master/LICENCE
-
-using McMaster.Extensions.CommandLineUtils;
-
-namespace PerformanceCalculator
-{
-    [HelpOption("-?|-h|--help")]
-    public abstract class CommandBase
-    {
-    }
-}
