--- PerformanceCalculator/Simulate/BaseSimulateCommand.cs
@@ -1,39 +1,0 @@
-// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
-// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-tools/master/LICENCE
-
-using JetBrains.Annotations;
-using osu.Game.Rulesets;
-
-namespace PerformanceCalculator.Simulate
-{
-    public abstract class BaseSimulateCommand : ProcessorCommand
-    {
-        public abstract string Beatmap { get; }
-
-        public abstract Ruleset Ruleset { get;  }
-
-        [UsedImplicitly]
-        public virtual double Accuracy { get; }
-
-        [UsedImplicitly]
-        public virtual int? Combo { get; }
-
-        [UsedImplicitly]
-        public virtual double PercentCombo { get; }
-
-        [UsedImplicitly]
-        public virtual int Score { get; }
-
-        [UsedImplicitly]
-        public virtual string[] Mods { get; }
-
-        [UsedImplicitly]
-        public virtual int Misses { get; }
-
-        [UsedImplicitly]
-        public virtual int? Mehs { get; }
-
-        [UsedImplicitly]
-        public virtual int? Goods { get; }
-    }
-}
