+++ PerformanceCalculator/Simulate/ManiaSimulateCommand.cs
@@ -2,15 +2,37 @@
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-tools/master/LICENCE
 
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Globalization;
+using JetBrains.Annotations;
+using McMaster.Extensions.CommandLineUtils;
 using osu.Game.Beatmaps;
+using osu.Game.Rulesets;
+using osu.Game.Rulesets.Mania;
 using osu.Game.Rulesets.Scoring;
 using osu.Game.Scoring;
 
-namespace PerformanceCalculator.Simulate.Mania
+namespace PerformanceCalculator.Simulate
 {
-    public class ManiaSimulateProcessor : BaseSimulateProcessor
+    [Command(Name = "simulate mania", Description = "Computes the performance (pp) of a simulated osu!mania play.")]
+    public class ManiaSimulateCommand : SimulateCommand
     {
+        [UsedImplicitly]
+        [Required, FileExists]
+        [Argument(0, Name = "beatmap", Description = "Required. The beatmap file (.osu).")]
+        public override string Beatmap { get; }
+
+        [UsedImplicitly]
+        [Option(Template = "-s|--score <score>", Description = "Score. An integer 0-1000000.")]
+        public override int Score { get; } = 1000000;
+
+        [UsedImplicitly]
+        [Option(CommandOptionType.MultipleValue, Template = "-m|--mod <mod>", Description = "One for each mod. The mods to compute the performance with."
+                                                                                            + " Values: hr, dt, fl, 4k, 5k, etc...")]
+        public override string[] Mods { get; }
+
+        public override Ruleset Ruleset => new ManiaRuleset();
+
         protected override int GetMaxCombo(IBeatmap beatmap) => 0;
 
         protected override Dictionary<HitResult, int> GenerateHitResults(double accuracy, IBeatmap beatmap, int countMiss, int? countMeh, int? countGood)
@@ -32,11 +54,6 @@ namespace PerformanceCalculator.Simulate.Mania
         protected override void WritePlayInfo(ScoreInfo scoreInfo, IBeatmap beatmap)
         {
             WriteAttribute("Score", scoreInfo.TotalScore.ToString(CultureInfo.InvariantCulture));
-        }
-
-        public ManiaSimulateProcessor(BaseSimulateCommand command)
-            : base(command)
-        {
         }
     }
 }
