+++ PerformanceCalculator/Simulate/TaikoSimulateCommand.cs
@@ -3,17 +3,57 @@
 
 using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Globalization;
 using System.Linq;
+using JetBrains.Annotations;
+using McMaster.Extensions.CommandLineUtils;
 using osu.Game.Beatmaps;
+using osu.Game.Rulesets;
 using osu.Game.Rulesets.Scoring;
+using osu.Game.Rulesets.Taiko;
 using osu.Game.Rulesets.Taiko.Objects;
 using osu.Game.Scoring;
 
-namespace PerformanceCalculator.Simulate.Taiko
+namespace PerformanceCalculator.Simulate
 {
-    public class TaikoSimulateProcessor : BaseSimulateProcessor
+    [Command(Name = "simulate taiko", Description = "Computes the performance (pp) of a simulated osu!taiko play.")]
+    public class TaikoSimulateCommand : SimulateCommand
     {
+        [UsedImplicitly]
+        [Required, FileExists]
+        [Argument(0, Name = "beatmap", Description = "Required. The beatmap file (.osu).")]
+        public override string Beatmap { get; }
+
+        [UsedImplicitly]
+        [Option(Template = "-a|--accuracy <accuracy>", Description = "Accuracy. Enter as decimal 0-100. Defaults to 100."
+                                                                     + " Scales hit results as well and is rounded to the nearest possible value for the beatmap.")]
+        public override double Accuracy { get; } = 100;
+
+        [UsedImplicitly]
+        [Option(Template = "-c|--combo <combo>", Description = "Maximum combo during play. Defaults to beatmap maximum.")]
+        public override int? Combo { get; }
+
+        [UsedImplicitly]
+        [Option(Template = "-C|--percent-combo <combo>", Description = "Percentage of beatmap maximum combo achieved. Alternative to combo option."
+                                                                       + " Enter as decimal 0-100.")]
+        public override double PercentCombo { get; } = 100;
+
+        [UsedImplicitly]
+        [Option(CommandOptionType.MultipleValue, Template = "-m|--mod <mod>", Description = "One for each mod. The mods to compute the performance with."
+                                                                                            + " Values: hr, dt, hd, fl, ez, etc...")]
+        public override string[] Mods { get; }
+
+        [UsedImplicitly]
+        [Option(Template = "-X|--misses <misses>", Description = "Number of misses. Defaults to 0.")]
+        public override int Misses { get; }
+
+        [UsedImplicitly]
+        [Option(Template = "-G|--goods <goods>", Description = "Number of goods. Will override accuracy if used. Otherwise is automatically calculated.")]
+        public override int? Goods { get; }
+
+        public override Ruleset Ruleset => new TaikoRuleset();
+
         protected override int GetMaxCombo(IBeatmap beatmap) => beatmap.HitObjects.OfType<Hit>().Count();
 
         protected override Dictionary<HitResult, int> GenerateHitResults(double accuracy, IBeatmap beatmap, int countMiss, int? countMeh, int? countGood)
@@ -61,11 +101,6 @@ namespace PerformanceCalculator.Simulate.Taiko
             WriteAttribute("Misses", scoreInfo.Statistics[HitResult.Miss].ToString(CultureInfo.InvariantCulture));
             WriteAttribute("Goods", scoreInfo.Statistics[HitResult.Good].ToString(CultureInfo.InvariantCulture));
             WriteAttribute("Greats", scoreInfo.Statistics[HitResult.Great].ToString(CultureInfo.InvariantCulture));
-        }
-
-        public TaikoSimulateProcessor(BaseSimulateCommand command)
-            : base(command)
-        {
         }
     }
 }
